Merge master branch with avr-gcc implementation of the code

+++ src/aes_share.h
@@ -7,6 +7,6 @@ void shiftrows_share(byte *stateshare[16],int n);
 void mixcolumns_share(byte *stateshare[16],int n);
 void addroundkey_share(byte *stateshare[16],byte *wshare[176],int round,int n);
 
-int run_aes_share(byte in[16],byte out[16],byte key[16],int n,void (*subbyte_share_call)(byte *,int),int nt);
+double run_aes_share(byte in[16],byte out[16],byte key[16],int n,void (*subbyte_share_call)(byte *,int),int nt);
 
 #endif
